Complete Project: Landing Page
Finalised the styling for headings, added a
transparent bg for the navbar and also made it
fixed. Added a drop shadow to the hero image for
some visual interest. Also added max widths to
the hero and features section as they didn't look
good on larger monitors. Need to learn the correct
way to handle that situation in future because I
just picked an arbitrary px width that looked
decent.

I probably went a bit overboard on this project,
but I thought it'd be a good place to start
building things related to my project, Modes.
Overall I'm really pleased with how the site
turned out. Obviously it isn't responsive yet,
that will be the next challenge.